refactor(resource-manager): move metrics-related code to metrics.go
