+++ pkg/mcs/resourcemanager/server/manager_test.go
@@ -167,6 +167,7 @@ func checkAddAndModifyResourceGroup(re *require.Assertions, manager *Manager, ke
 
 func TestCleanUpTicker(t *testing.T) {
 	re := require.New(t)
+	initMaps()
 	m := prepareManager()
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -175,12 +176,12 @@ func TestCleanUpTicker(t *testing.T) {
 	keyspaceID := uint32(1)
 	prepareKeyspaceName(ctx, re, m, &rmpb.KeyspaceIDValue{Value: keyspaceID}, "test_keyspace")
 	// Insert two consumption records manually.
-	m.consumptionRecord[consumptionRecordKey{
+	consumptionRecordMap[consumptionRecordKey{
 		keyspaceID: keyspaceID,
 		groupName:  "test_group_1",
 		ruType:     defaultTypeLabel,
 	}] = time.Now().Add(-metricsCleanupTimeout * 2)
-	m.consumptionRecord[consumptionRecordKey{
+	consumptionRecordMap[consumptionRecordKey{
 		keyspaceID: keyspaceID,
 		groupName:  "test_group_2",
 		ruType:     defaultTypeLabel,
@@ -197,8 +198,8 @@ func TestCleanUpTicker(t *testing.T) {
 	// Close the manager to avoid the data race.
 	m.close()
 
-	re.Len(m.consumptionRecord, 1)
-	re.Contains(m.consumptionRecord, consumptionRecordKey{
+	re.Len(consumptionRecordMap, 1)
+	re.Contains(consumptionRecordMap, consumptionRecordKey{
 		keyspaceID: keyspaceID,
 		groupName:  "test_group_2",
 		ruType:     defaultTypeLabel,
